requirements.txt
torch==2.2.1
torchvision==0.17.1
torchaudio==2.2.1
scikit-learn
opencv-python
numpy
pandas
matplotlib
black
flake8
jupyterlab
jupyterlab-code-formatter
tqdm
gdown
nbstripout
huggingface_hub
transformers
python-dotenv
roboflow
tensorboard
boto3
joblib
pillow